pyproject.toml
[project]
name = "TruckRoute Pro"        
version = "0.1.0"
description = "Django backend for Trip Route Log"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.6",
    "django-cors-headers>=4.8.0",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework>=3.16.1",
    "psycopg2-binary>=2.9.10",
    "python-decouple>=3.8",
    "requests>=2.32.5",
]
